# Refactor: pairings