--- a/ecc/bls12-377/pairing.go
+++ b/ecc/bls12-377/pairing.go
@@ -133,12 +133,16 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 	}
 
 	var result GT
-	result.SetOne()
-
 	var l lineEvaluation
 
 	// i == len(loopCounter) - 2
-	for k := 0; k < n; k++ {
+	qProj[0].doubleStep(&l)
+	// line eval
+	result.C0.B0.MulByElement(&l.r0, &p[0].Y)
+	result.C1.B0.MulByElement(&l.r1, &p[0].X)
+	result.C1.B1.Set(&l.r2)
+
+	for k := 1; k < n; k++ {
 		qProj[k].doubleStep(&l)
 		// line eval
 		l.r0.MulByElement(&l.r0, &p[k].Y)