--- a/ecc/bls12-377/pairing.go
+++ b/ecc/bls12-377/pairing.go
@@ -104,7 +104,7 @@ func FinalExponentiation(z *GT, _z ...*GT) GT {
 }
 
 // MillerLoop computes the multi-Miller loop
-// ∏ᵢ MillerLoop(Pᵢ, Qᵢ)
+// ∏ᵢ MillerLoop(Pᵢ, Qᵢ) = ∏ᵢ { fᵢ_{x,Qᵢ}(Pᵢ) }
 func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 	// check input size match
 	n := len(P)
@@ -136,19 +136,30 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 	var l1, l2 lineEvaluation
 	var prodLines [5]E2
 
-	// i == len(loopCounter) - 2
+	// Compute ∏ᵢ { fᵢ_{x₀,Q}(P) }
+	// i = 62, separately to avoid an E12 Square
+	// (Square(res) = 1² = 1)
+	// loopCounter[62] = 0
+	// k = 0, separately to avoid MulBy034 (res × ℓ)
+	// (assign line to res)
+
+	// qProj[0] ← 2qProj[0] and l1 the tangent ℓ passing 2qProj[0]
 	qProj[0].doubleStep(&l1)
-	// line eval
+	// line evaluation at P[0] (assign)
 	result.C0.B0.MulByElement(&l1.r0, &p[0].Y)
 	result.C1.B0.MulByElement(&l1.r1, &p[0].X)
 	result.C1.B1.Set(&l1.r2)
 
 	if n >= 2 {
-		// k = 1
+		// k = 1, separately to avoid MulBy034 (res × ℓ)
+		// (res is also a line at this point, so we use Mul034By034 ℓ × ℓ)
+
+		// qProj[1] ← 2qProj[1] and l1 the tangent ℓ passing 2qProj[1]
 		qProj[1].doubleStep(&l1)
-		// line evaluation
+		// line evaluation at P[1]
 		l1.r0.MulByElement(&l1.r0, &p[1].Y)
 		l1.r1.MulByElement(&l1.r1, &p[1].X)
+		// ℓ × res
 		prodLines = fptower.Mul034By034(&l1.r0, &l1.r1, &l1.r2, &result.C0.B0, &result.C1.B0, &result.C1.B1)
 		result.C0.B0 = prodLines[0]
 		result.C0.B1 = prodLines[1]
@@ -157,52 +168,67 @@ func MillerLoop(P []G1Affine, Q []G2Affine) (GT, error) {
 		result.C1.B1 = prodLines[4]
 	}
 
+	// k >= 2
 	for k := 2; k < n; k++ {
+		// qProj[k] ← 2qProj[k] and l1 the tangent ℓ passing 2qProj[k]
 		qProj[k].doubleStep(&l1)
-		// line eval
+		// line evaluation at P[k]
 		l1.r0.MulByElement(&l1.r0, &p[k].Y)
 		l1.r1.MulByElement(&l1.r1, &p[k].X)
+		// ℓ × res
 		result.MulBy034(&l1.r0, &l1.r1, &l1.r2)
 	}
 
+	// i <= 61
 	for i := len(loopCounter) - 3; i >= 1; i-- {
+		// mutualize the square among n Miller loops
 		// (∏ᵢfᵢ)²
 		result.Square(&result)
 
 		for k := 0; k < n; k++ {
+			// qProj[k] ← 2qProj[k] and l1 the tangent ℓ passing 2qProj[k]
 			qProj[k].doubleStep(&l1)
-			// line eval
+			// line evaluation at P[k]
 			l1.r0.MulByElement(&l1.r0, &p[k].Y)
 			l1.r1.MulByElement(&l1.r1, &p[k].X)
 
 			if loopCounter[i] == 0 {
+				// ℓ × res
 				result.MulBy034(&l1.r0, &l1.r1, &l1.r2)
 			} else {
+				// qProj[k] ← qProj[k]+Q[k] and
+				// l2 the line ℓ passing qProj[k] and Q[k]
 				qProj[k].addMixedStep(&l2, &q[k])
-				// line eval
+				// line evaluation at P[k]
 				l2.r0.MulByElement(&l2.r0, &p[k].Y)
 				l2.r1.MulByElement(&l2.r1, &p[k].X)
+				// ℓ × ℓ
 				prodLines = fptower.Mul034By034(&l1.r0, &l1.r1, &l1.r2, &l2.r0, &l2.r1, &l2.r2)
+				// (ℓ × ℓ) × res
 				result.MulBy01234(&prodLines)
 			}
 		}
 
 	}
 
-	// i = 0
+	// i = 0, separately to avoid a point addition
+	// loopCounter[0] = 1
 	result.Square(&result)
 	for k := 0; k < n; k++ {
+		// qProj[k] ← 2qProj[k] and l1 the tangent ℓ passing 2qProj[k]
 		qProj[k].doubleStep(&l1)
-		// line eval
+		// line evaluation at P[k]
 		l1.r0.MulByElement(&l1.r0, &p[k].Y)
 		l1.r1.MulByElement(&l1.r1, &p[k].X)
 
+		// l2 the line passing qProj[k] and Q
 		qProj[k].lineCompute(&l2, &q[k])
-		// line eval
+		// line evaluation at P[k]
 		l2.r0.MulByElement(&l2.r0, &p[k].Y)
 		l2.r1.MulByElement(&l2.r1, &p[k].X)
-
+		// ℓ × ℓ
 		prodLines = fptower.Mul034By034(&l1.r0, &l1.r1, &l1.r2, &l2.r0, &l2.r1, &l2.r2)
+		// (ℓ × ℓ) × res
 		result.MulBy01234(&prodLines)
 	}