--- a/ecc/bls12-381/internal/fptower/e12_pairing.go
+++ b/ecc/bls12-381/internal/fptower/e12_pairing.go
@@ -67,8 +67,8 @@ func (z *E12) MulBy014(c0, c1, c4 *E2) *E12 {
 }
 
 // Mul014By014 multiplication of sparse element (c0,c1,0,0,c4,0) by sparse element (d0,d1,0,0,d4,0)
-func (z *E12) Mul014By014(d0, d1, d4, c0, c1, c4 *E2) *E12 {
-	var tmp, x0, x1, x4, x04, x01, x14 E2
+func Mul014By014(d0, d1, d4, c0, c1, c4 *E2) [5]E2 {
+	var z00, tmp, x0, x1, x4, x04, x01, x14 E2
 	x0.Mul(c0, d0)
 	x1.Mul(c1, d1)
 	x4.Mul(c4, d4)
@@ -88,13 +88,30 @@ func (z *E12) Mul014By014(d0, d1, d4, c0, c1, c4 *E2) *E12 {
 		Sub(&x14, &x1).
 		Sub(&x14, &x4)
 
-	z.C0.B0.MulByNonResidue(&x4).
-		Add(&z.C0.B0, &x0)
-	z.C0.B1.Set(&x01)
-	z.C0.B2.Set(&x1)
-	z.C1.B0.SetZero()
-	z.C1.B1.Set(&x04)
-	z.C1.B2.Set(&x14)
+	z00.MulByNonResidue(&x4).
+		Add(&z00, &x0)
+
+	return [5]E2{z00, x01, x1, x04, x14}
+}
+
+// MulBy01245 multiplies z by an E12 sparse element of the form (x0, x1, x2, 0, x4, x5)
+func (z *E12) MulBy01245(x *[5]E2) *E12 {
+	var c1, a, b, c, z0, z1 E6
+	c0 := &E6{B0: x[0], B1: x[1], B2: x[2]}
+	c1.B1 = x[3]
+	c1.B2 = x[4]
+	a.Add(&z.C0, &z.C1)
+	b.Add(c0, &c1)
+	a.Mul(&a, &b)
+	b.Mul(&z.C0, c0)
+	c.Set(&z.C1).MulBy12(&x[3], &x[4])
+	z1.Sub(&a, &b)
+	z1.Sub(&z1, &c)
+	z0.MulByNonResidue(&c)
+	z0.Add(&z0, &b)
+
+	z.C0 = z0
+	z.C1 = z1
 
 	return z
 }