Some refactoring and small optimisations of pairings ported from gnark.