@yelhousni 2023-03-26
Some benchmarks on `Intel(R) Xeon(R) Platinum 8151 CPU @ 3.40GHz`.

BN254:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkPairing-2                411813        406969        -1.18%
BenchmarkMultiPair/2_pairs-2      555552        548078        -1.35%
BenchmarkMultiPair/3_pairs-2      697761        688646        -1.31%
```
BLS12-381:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkPairing-2                627514        623349        -0.66%
BenchmarkMultiPair/2_pairs-2      819556        815842        -0.45%
BenchmarkMultiPair/3_pairs-2      1010009       1005709       -0.43%
```
BLS12-377:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkPairing-2                847109        841373        -0.68%
BenchmarkMultiPair/2_pairs-2      1098540       1085993       -1.14%
BenchmarkMultiPair/3_pairs-2      1343194       1325140       -1.34%
```
BW6-761:
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkPairing-2                2257147       2247891       -0.41%
BenchmarkMultiPair/2_pairs-2      3034059       3025737       -0.27%
BenchmarkMultiPair/3_pairs-2      3802704       3785700       -0.45%
```